Rename check-nilaway-func-size testdata package to example
Matches the convention already used by the repo's other two custom
analyzers, tools/ci/setboolcheck and tools/ci/apiparamcheck, which both
put their fixtures in testdata/src/example/example.go.

Git tracks this as a pure rename; the only content changes are the
package clause and the package pattern passed to analysistest.Run.
